+++ src/engine.rs
@@ -519,8 +519,8 @@ pub struct State {
     /// Embedded module resolver.
     #[cfg(not(feature = "no_module"))]
     pub resolver: Option<Shared<crate::module::resolvers::StaticModuleResolver>>,
-    /// Cached lookup values for function hashes.
-    pub functions_caches: StaticVec<
+    /// Functions resolution cache.
+    fn_resolution_caches: StaticVec<
         HashMap<
             NonZeroU64,
             Option<(CallableFunction, Option<ImmutableString>)>,
@@ -534,6 +534,48 @@ impl State {
     #[inline(always)]
     pub fn is_global(&self) -> bool {
         self.scope_level == 0
+    }
+    /// Get the current functions resolution cache.
+    pub fn fn_resolution_cache(
+        &self,
+    ) -> Option<
+        &HashMap<
+            NonZeroU64,
+            Option<(CallableFunction, Option<ImmutableString>)>,
+            StraightHasherBuilder,
+        >,
+    > {
+        self.fn_resolution_caches.last()
+    }
+    /// Get a mutable reference to the current functions resolution cache.
+    pub fn fn_resolution_cache_mut(
+        &mut self,
+    ) -> &mut HashMap<
+        NonZeroU64,
+        Option<(CallableFunction, Option<ImmutableString>)>,
+        StraightHasherBuilder,
+    > {
+        if self.fn_resolution_caches.is_empty() {
+            self.fn_resolution_caches
+                .push(HashMap::with_capacity_and_hasher(16, StraightHasherBuilder));
+        }
+        self.fn_resolution_caches.last_mut().unwrap()
+    }
+    /// Push an empty functions resolution cache onto the stack and make it current.
+    pub fn push_fn_resolution_cache(&mut self) {
+        self.fn_resolution_caches.push(Default::default());
+    }
+    /// Remove the current functions resolution cache and make the last one current.
+    pub fn pop_fn_resolution_cache(&mut self) {
+        self.fn_resolution_caches.pop();
+    }
+    /// Clear the current functions resolution cache.
+    ///
+    /// # Panics
+    ///
+    /// Panics if there is no current functions resolution cache.
+    pub fn clear_fn_resolution_cache(&mut self) {
+        self.fn_resolution_caches.last_mut().unwrap().clear();
     }
 }
 
@@ -1712,7 +1754,7 @@ impl Engine {
 
             // Statement block
             Expr::Stmt(x, _) => {
-                self.eval_stmt_block(scope, mods, state, lib, this_ptr, x.as_ref(), level)
+                self.eval_stmt_block(scope, mods, state, lib, this_ptr, x.as_ref(), true, level)
             }
 
             // lhs[idx_expr]
@@ -1856,45 +1898,49 @@ impl Engine {
         lib: &[&Module],
         this_ptr: &mut Option<&mut Dynamic>,
         statements: impl IntoIterator<Item = &'a Stmt>,
+        restore: bool,
         level: usize,
     ) -> Result<Dynamic, Box<EvalAltResult>> {
-        let mut has_imports = false;
+        let mut _has_imports = false;
         let prev_always_search = state.always_search;
         let prev_scope_len = scope.len();
         let prev_mods_len = mods.len();
-        state.scope_level += 1;
 
-        let result = statements
-            .into_iter()
-            .try_fold(Default::default(), |_, stmt| {
-                match stmt {
-                    #[cfg(not(feature = "no_module"))]
-                    Stmt::Import(_, _, _) => {
-                        // When imports list is modified, clear the functions lookup cache
-                        if has_imports {
-                            state.functions_caches.last_mut().map(|c| c.clear());
-                        } else {
-                            state.functions_caches.push(Default::default());
-                        }
-                        has_imports = true;
-                    }
-                    _ => (),
-                }
-
-                self.eval_stmt(scope, mods, state, lib, this_ptr, stmt, level)
-            });
-
-        scope.rewind(prev_scope_len);
-        if has_imports {
-            // If imports list is modified, pop the functions lookup cache
-            state.functions_caches.pop();
+        if restore {
+            state.scope_level += 1;
         }
-        mods.truncate(prev_mods_len);
-        state.scope_level -= 1;
 
-        // The impact of new local variables goes away at the end of a block
-        // because any new variables introduced will go out of scope
-        state.always_search = prev_always_search;
+        let result = statements.into_iter().try_fold(Dynamic::UNIT, |_, stmt| {
+            #[cfg(not(feature = "no_module"))]
+            match stmt {
+                Stmt::Import(_, _, _) => {
+                    // When imports list is modified, clear the functions lookup cache
+                    if _has_imports {
+                        state.clear_fn_resolution_cache();
+                    } else if restore {
+                        state.push_fn_resolution_cache();
+                        _has_imports = true;
+                    }
+                }
+                _ => (),
+            }
+
+            self.eval_stmt(scope, mods, state, lib, this_ptr, stmt, level)
+        });
+
+        if restore {
+            scope.rewind(prev_scope_len);
+            if _has_imports {
+                // If imports list is modified, pop the functions lookup cache
+                state.pop_fn_resolution_cache();
+            }
+            mods.truncate(prev_mods_len);
+            state.scope_level -= 1;
+
+            // The impact of new local variables goes away at the end of a block
+            // because any new variables introduced will go out of scope
+            state.always_search = prev_always_search;
+        }
 
         result
     }
@@ -2088,7 +2134,7 @@ impl Engine {
 
             // Block scope
             Stmt::Block(statements, _) => {
-                self.eval_stmt_block(scope, mods, state, lib, this_ptr, statements, level)
+                self.eval_stmt_block(scope, mods, state, lib, this_ptr, statements, true, level)
             }
 
             // If statement
@@ -2246,6 +2292,7 @@ impl Engine {
 
                 match result {
                     Ok(_) => result,
+                    Err(err) if err.is_pseudo_error() => Err(err),
                     Err(err) if !err.is_catchable() => Err(err),
                     Err(mut err) => {
                         let value = match *err {