+++ src/engine_api.rs
@@ -1527,13 +1527,16 @@ impl Engine {
         ast: &'a AST,
         level: usize,
     ) -> Result<Dynamic, Box<EvalAltResult>> {
-        let state = &mut State {
-            source: ast.clone_source(),
-            #[cfg(not(feature = "no_module"))]
-            resolver: ast.resolver(),
-            ..Default::default()
-        };
-        self.eval_statements_raw(scope, mods, state, ast.statements(), &[ast.lib()], level)
+        let mut state: State = Default::default();
+        state.source = ast.clone_source();
+        #[cfg(not(feature = "no_module"))]
+        {
+            state.resolver = ast.resolver();
+        }
+
+        let statements = ast.statements();
+        let lib = &[ast.lib()];
+        self.eval_global_statements(scope, mods, &mut state, statements, lib, level)
     }
     /// Evaluate a file, but throw away the result and only return error (if any).
     /// Useful for when you don't need the result, but still need to keep track of possible errors.
@@ -1596,13 +1599,15 @@ impl Engine {
         ast: &AST,
     ) -> Result<(), Box<EvalAltResult>> {
         let mods = &mut (&self.global_sub_modules).into();
-        let state = &mut State {
-            source: ast.clone_source(),
-            #[cfg(not(feature = "no_module"))]
-            resolver: ast.resolver(),
-            ..Default::default()
-        };
-        self.eval_statements_raw(scope, mods, state, ast.statements(), &[ast.lib()], 0)?;
+        let mut state: State = Default::default();
+        state.source = ast.clone_source();
+        #[cfg(not(feature = "no_module"))]
+        {
+            state.resolver = ast.resolver();
+        }
+        let statements = ast.statements();
+        let lib = &[ast.lib()];
+        self.eval_global_statements(scope, mods, &mut state, statements, lib, 0)?;
         Ok(())
     }
     /// Call a script function defined in an [`AST`] with multiple arguments.